Refactor Sura model and enhance review completion: update page count to double, improve completion dialog with reviewed pages information and update database schema
- Modify `Sura` model to use `double` for `pages` property to accommodate partial page counts.
- Update database schema to reflect the change in the `pages` column type to `REAL` in the `suras` and `selected_suras` tables, and `completed_pages` column in `daily_progress`.
- In `HomeScreen`, modify progress calculation to handle `double` values and add the reviewed pages count in the completion dialog.
- In `SelectSurasScreen`, update dialogs to use `double` for page inputs and modify database interaction accordingly.
- Enhance `reviewCompletedWithPages` translation messages to show the reviewed pages count in Arabic and English.

+++ lib/models/sura_model.dart
@@ -1,7 +1,7 @@
 class Sura {
   final int id;
   final String name;
-  final int pages;
+  final double pages;
   bool isCompleted;
 
   Sura({
